feat(receive): share receive reducer
