+++ packages/suite/package.json
@@ -51,6 +51,7 @@
         "@suite-common/mev": "workspace:*",
         "@suite-common/platform-encryption": "workspace:*",
         "@suite-common/react-query": "workspace:*",
+        "@suite-common/receive": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/sentry": "workspace:*",
         "@suite-common/staking": "workspace:*",
