+++ packages/suite/src/actions/suite/__tests__/storageActions.test.ts
@@ -3,11 +3,11 @@ import '@suite-common/test-utils/src/globalOverrides';
 import { coinjoinReducer } from '@suite/coinjoin';
 import { initialRunCompleted, prepareFlagsReducer } from '@suite/flags';
 import { initialMetadataState, metadataReducer } from '@suite/metadata';
-import { receiveReducer } from '@suite/receive';
 import { suiteSettingsInitialState } from '@suite/settings';
 import { suiteSyncSlice } from '@suite/suite-sync';
 import { deviceActions, selectDevices, selectDevicesCount } from '@suite-common/device';
 import { asEncryptedHex } from '@suite-common/platform-encryption';
+import { receiveReducer } from '@suite-common/receive';
 import { setSuiteSyncOwner } from '@suite-common/suite-sync';
 import { type SuiteSyncOwnerSerialized } from '@suite-common/suite-sync-storage';
 import { mockSuiteDevice } from '@suite-common/suite-types/mocks';
@@ -113,7 +113,6 @@ type PartialState = Pick<
             | 'settings'
             | 'discovery'
             | 'send'
-            | 'receive'
             | 'transactions'
             | 'graph'
             | 'fiat'
@@ -147,6 +146,7 @@ const getInitialState = (prevState?: Partial<PartialState>, action?: any) => ({
         prevState ? prevState.device : undefined,
         action || ({ type: 'foo' } as any),
     ),
+    receive: receiveReducer(prevState?.receive, action || ({ type: 'foo' } as any)),
     wallet: {
         accounts: accountsReducer(prevState?.wallet?.accounts, action || ({ type: 'foo' } as any)),
         coinjoin: coinjoinReducer(prevState?.wallet?.coinjoin, action || ({ type: 'foo' } as any)),
@@ -159,7 +159,6 @@ const getInitialState = (prevState?: Partial<PartialState>, action?: any) => ({
             action || ({ type: 'foo' } as any),
         ),
         send: sendFormReducer(prevState?.wallet?.send, action || ({ type: 'foo' } as any)),
-        receive: receiveReducer(prevState?.wallet?.receive, action || ({ type: 'foo' } as any)),
         transactions: transactionsReducer(
             prevState?.wallet?.transactions,
             action || ({ type: 'foo' } as any),
