+++ suite/receive/src/index.ts
@@ -1,13 +1,3 @@
-export {
-    receiveActions,
-    receiveReducer,
-    selectCurrentFreshAddress,
-    selectReceiveRevealedAddresses,
-    type CurrentFreshAddress,
-    type ReceiveAccountState,
-    type ReceiveRootState,
-    type ReceiveState,
-} from './receiveReducer';
 export { CoinQrCode } from './CoinQrCode';
 export { FreshAddress, type FreshAddressProps } from './FreshAddress';
 export { openAddressModal } from './openAddressModal';
